+++ cypress/e2e/group-1/source_genome_region.cy.js
@@ -1,4 +1,4 @@
-import { clearInputValue, clickMultiSelectOption, setInputValue, skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { clearInputValue, clickMultiSelectOption, setInputValue, skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('GenomeRegion Source', () => {
   beforeEach(() => {
