+++ cypress/e2e/group-1/source_polymerase_extension.cy.js
@@ -1,4 +1,4 @@
-import { addLane, addSource, clickMultiSelectOption, manuallyTypeSequence, skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { addLane, addSource, clickMultiSelectOption, manuallyTypeSequence, skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Test polymerase extension functionality', () => {
   beforeEach(() => {
