+++ cypress/e2e/group-3/source_empty.cy.js
@@ -1,4 +1,4 @@
-import { skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Test empty source functionality', () => {
   beforeEach(() => {
