+++ cypress/e2e/group-4/download_history_file.cy.js
@@ -1,5 +1,5 @@
 import { BlobReader, TextWriter, ZipReader } from '@zip.js/zip.js';
-import { skipGoogleSheetErrors, skipNcbiCheck, setInputValue, deleteSourceByContent, loadExample } from './common_functions';
+import { skipGoogleSheetErrors, skipNcbiCheck, setInputValue, deleteSourceByContent, loadExample } from '../common_functions';
 
 describe('Test download history file', () => {
   beforeEach(() => {
