+++ cypress/e2e/group-4/download_sequence_file.cy.js
@@ -1,4 +1,4 @@
-import { setInputValue, skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { setInputValue, skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Test download sequence file', () => {
   beforeEach(() => {
