+++ cypress/e2e/group-4/drag_and_drop_file.cy.js
@@ -1,4 +1,4 @@
-import { loadExample, skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { loadExample, skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Test drag and drop functionality', () => {
   beforeEach(() => {
