+++ cypress/e2e/group-4/load_example.cy.js
@@ -1,4 +1,4 @@
-import { skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Test load example functionality', () => {
   beforeEach(() => {
