+++ cypress/e2e/group-4/make_examples.cy.js
@@ -1,4 +1,4 @@
-import { addSource, clickMultiSelectOption, setInputValue, addPrimer, addLane, skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { addSource, clickMultiSelectOption, setInputValue, addPrimer, addLane, skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Makes all examples', () => {
   beforeEach(() => {
