+++ cypress/e2e/group-4/tab_navigation.cy.js
@@ -1,4 +1,4 @@
-import { skipGoogleSheetErrors, skipNcbiCheck } from './common_functions';
+import { skipGoogleSheetErrors, skipNcbiCheck } from '../common_functions';
 
 describe('Test tab navigation functionality', () => {
   beforeEach(() => {
